refactor: MapStruct 불필요한 옵션 제거 및 디렉토리명을 mapper에서 mapstruct으로 변경 (#72)
