+++ src/main/java/com/harusari/chainware/franchise/common/mapstruct/FranchiseMapStruct.java
@@ -1,14 +1,10 @@
-package com.harusari.chainware.franchise.common.mapper;
+package com.harusari.chainware.franchise.common.mapstruct;
 
 import com.harusari.chainware.franchise.command.domain.aggregate.Franchise;
 import com.harusari.chainware.member.command.application.dto.request.franchise.FranchiseCreateRequest;
 import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
 
-@Mapper(
-        unmappedSourcePolicy =  ReportingPolicy.WARN,
-        unmappedTargetPolicy = ReportingPolicy.WARN
-)
+@Mapper
 public interface FranchiseMapStruct {
 
     Franchise toFranchise(FranchiseCreateRequest franchiseCreateRequest, Long memberId);
