+++ src/main/java/com/harusari/chainware/member/common/mapstruct/MemberMapStruct.java
@@ -1,14 +1,10 @@
-package com.harusari.chainware.member.common.mapper;
+package com.harusari.chainware.member.common.mapstruct;
 
 import com.harusari.chainware.member.command.application.dto.request.MemberCreateRequest;
 import com.harusari.chainware.member.command.domain.aggregate.Member;
 import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
 
-@Mapper(
-        unmappedSourcePolicy =  ReportingPolicy.WARN,
-        unmappedTargetPolicy = ReportingPolicy.WARN
-)
+@Mapper
 public interface MemberMapStruct {
 
     Member toMember(MemberCreateRequest memberCreateRequest);
