+++ src/main/java/com/harusari/chainware/warehouse/common/mapstruct/WarehouseMapStruct.java
@@ -1,14 +1,10 @@
-package com.harusari.chainware.warehouse.common.mapper;
+package com.harusari.chainware.warehouse.common.mapstruct;
 
 import com.harusari.chainware.member.command.application.dto.request.warehouse.WarehouseCreateRequest;
 import com.harusari.chainware.warehouse.command.domain.aggregate.Warehouse;
 import org.mapstruct.Mapper;
-import org.mapstruct.ReportingPolicy;
 
-@Mapper(
-        unmappedSourcePolicy =  ReportingPolicy.WARN,
-        unmappedTargetPolicy = ReportingPolicy.WARN
-)
+@Mapper
 public interface WarehouseMapStruct {
 
     Warehouse toWarehouse(WarehouseCreateRequest warehouseCreateRequest, Long memberId);
