+++ src/main/resources/mappers/warehouse/WarehouseInventorySnapshotQueryMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.harusari.chainware.warehouse.common.mapper.WarehouseInventorySnapshotQueryMapper">
+<mapper namespace="com.harusari.chainware.warehouse.command.mapper.WarehouseInventorySnapshotQueryMapper">
 
     <select id="findAllCurrentInventory" resultType="com.harusari.chainware.warehouse.command.application.dto.CurrentWarehouseInventoryDto">
         SELECT
